Title: 물의 상태, Time: 40ms, Memory: 16MB, Status: Wrong Answer - Codetree

+++ 250328/물의 상태/state-of-water.py
@@ -1,2 +1,2 @@
-n = int(input())
+n = float(input())
 print('ice' if n<0 else 'vapor' if n>100 else 'water');
